+++ experimental/canvaskit/CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+ - `SkFont` now exposed.
+
+### Removed
+- `SkPaint.measureText` - use `SkFont.measureText` instead.
 
 ## [0.3.0] - 2018-12-18